style(#24): rename page with lower camel case
