+++ src/page/about/index.tsx
@@ -3,14 +3,14 @@ import {
   useLocation
 } from 'react-router-dom';
 
-import { aboutHeader } from "../Config/About/AboutData";
+import { aboutHeader } from "../../Config/About/AboutData";
 
-import NavBar from "../components/Navbar";
-import Sidebar from "../components/SideBar/SideBar";
-import AboutText from "../components/About/AboutText";
-import LifeStyles from "../components/About/LifeStyles";
-import Github from "../components/About/Github";
-import TechStack from "../components/About/TechStack";
+import NavBar from "../../components/Navbar";
+import Sidebar from "../../components/SideBar/SideBar";
+import AboutText from "../../components/About/AboutText";
+import LifeStyles from "../../components/About/LifeStyles";
+import Github from "../../components/About/Github";
+import TechStack from "../../components/About/TechStack";
 // import Clients from "../components/About/Clients";
 // import Testimonial from "../components/About/Testimonial";
